requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
click==8.2.1
colorama==0.4.6
distro==1.9.0
eval_type_backport==0.2.2
fastapi==0.115.12
griffe==1.7.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
jiter==0.10.0
logfire-api==3.19.0
openai==1.86.0
opentelemetry-api==1.34.1
pydantic==2.11.6
pydantic-ai-slim==0.2.18
pydantic-graph==0.2.18
pydantic-settings==2.9.1
pydantic_core==2.33.2
python-dotenv==1.1.0
sniffio==1.3.1
starlette==0.46.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.34.3
zipp==3.23.0
